refactor, remove robots
