+++ CMakeLists.txt
@@ -94,7 +94,7 @@ set(CMAKE_SHARED_LINKER_FLAGS "${CMAKE_SHARED_LINKER_FLAGS} -fuse-ld=lld -Wl,--d
 
 # Examples
 if(KALMAN_BUILD_EXAMPLES)
-  add_executable(example_robot1 "examples/Robot1/big_try.cpp")
+  add_executable(example_robot1 "examples/big/big_try.cpp")
   #target_compile_options(example_robot1 PUBLIC)
   #add_link_options(-mllvm -enzyme-print -enzyme-loose-types=1)
   target_link_libraries(example_robot1 PUBLIC LLDEnzymeFlags ${BLAS_LIBRARIES})
